Extract thread message utilities
